requirements.txt
evadb[document]
openai
instructor
pydantic==2.4.0
python-dotenv==1.0.0
tiktoken